crimson/mon: move mon::Connection into .cc
